R3a: flip import direction + swap E-layer into consumer call sites
Finding: metric_estimators imported _spd_mean/_relative_pd_floor from
low_rank_adaptation, creating an inverted dependency (the E-layer
depending on a consumer).  meads_adaptation._lrd_from_accumulated_covariance
and mclmc_lrd_adaptation._extract_lrd_from_samples had inline math
identical to E-layer functions.  mclmc/adjusted_mclmc streaming variance
sites are NOT expressible as an array-call because x_average/x_squared_average
are incremental-value-update aggregates with no raw draws buffer at the
call site.

Fix:
- Dependency flip (swap 2a): move _spd_mean + _relative_pd_floor into
  metric_estimators; low_rank_adaptation now imports them from there.
  _compute_low_rank_metric body unchanged (n is traced JAX integer;
  buffer masking integral to all computation — see added docstring note).
- Swap 4 (meads): _lrd_from_accumulated_covariance body → one-liner
  calling sample_covariance_eigh_low_rank(acc.m2, acc.count, k).
  -8 LOC, behavior-identical (golden tests: 117/117 green).
- Swap 3 (mclmc_lrd): _extract_lrd_from_samples adopts E-layer utility
  select_top_eigenvalues_by_informativeness for the top-k step while
  keeping the single SVD pass. Full draws_singular_value_low_rank would
  require a second SVD to produce lam_all_sorted (G-layer diagnostic
  output not in the E-layer return type) — kept partial to avoid
  redundant SVD, per "do not contort" constraint.
- Swap 1 (mclmc + adjusted_mclmc): skipped with cross-reference comments
  at both inline variance sites (streaming aggregates, no raw draws).
