+++ blackjax/adaptation/adjusted_mclmc_adaptation.py
@@ -372,6 +372,12 @@ def adjusted_mclmc_make_L_step_size_adaptation(
         eigenvector = None
         if num_steps2 != 0.0:
             x_average, x_squared_average = average[0], average[1]
+            # R3a E-layer cross-reference: metric_estimators.sample_variance_diagonal
+            # implements the same formula (E[x^2] - E[x]^2) on raw draws.
+            # Body swap SKIPPED: x_average/x_squared_average are streaming
+            # incremental-value-update aggregates (not a raw draws array), so
+            # calling sample_variance_diagonal(draws) would require materializing
+            # the full draws buffer — state not available at this call site.
             variances = x_squared_average - jnp.square(x_average)
 
             if max == "max":
